Rename the library files to 3.0. (GH-1)
The philosophy betwwen 2.0 and 3.0 is a little bit different. WhoLib
version 2 tended to handle the /who & usual C_FriendList.SendWho API on
our own (i.e., queuing them).

The version 3 changed this behavior, and made them much like what they
will get treated as if they are calling the API themselves.

Therefore, you will get 2 type of API now, one is the original WOW API,
and the other is the WhoLib one, which is guaranteed to return a result.